3.8 en proceso

+++ src/main/java/org/iesvdm/dto/ComercialDTO2.java
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.iesvdm.modelo.Comercial;
 
+import java.math.BigDecimal;
+
 @Data
 public class ComercialDTO2 extends Comercial{
 
@@ -15,7 +17,7 @@ public class ComercialDTO2 extends Comercial{
     private int pedidosUltimoAnio;
     private int pedidosUltimoLustro;
 
-    public ComercialDTO2(int id, String nombre, String apellido1, String apellido2, float comision,
+    public ComercialDTO2(int id, String nombre, String apellido1, String apellido2, BigDecimal comision,
                          int numeroPedidosPorCliente, int pedidosUltimoTrimestre,
                          int pedidosUltimoSemestre, int pedidosUltimoAnio, int pedidosUltimoLustro) {
         super(id, nombre, apellido1, apellido2, comision);
